`platform_pki_2_electric_boogaloo` Windows implementation: re-use `schannel_utils.rs`
- Remove no longer needed `PrivateKey` & `acquire_private_key`
